+++ src/Evolution/DiscontinuousGalerkin/CMakeLists.txt
@@ -1,4 +1,4 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-add_subdirectory(SlopeLimiters)
+add_subdirectory(Limiters)
