+++ src/Evolution/DiscontinuousGalerkin/Limiters/WenoHelpers.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Evolution/DiscontinuousGalerkin/SlopeLimiters/WenoHelpers.hpp"
+#include "Evolution/DiscontinuousGalerkin/Limiters/WenoHelpers.hpp"
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/IndexIterator.hpp"  // IWYU pragma: keep
